Add files via upload
Welcome to the Netflix Clone HTML/CSS repository! This project is designed to provide you with a simplified version of the popular streaming platform Netflix, reimagined using HTML and CSS. Whether you're a beginner looking to learn web development or an experienced coder exploring new techniques, this repository is a valuable resource.